TimedMemberStateFactory periodic task checking partition safety should be scheduled on ManagementCenterService.start() instead of constructor.
