+++ hazelcast/src/main/java/com/hazelcast/management/TimedMemberStateFactoryHelper.java
@@ -1,3 +1,19 @@
+/*
+ * Copyright (c) 2008-2014, Hazelcast, Inc. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package com.hazelcast.management;
 
 import com.hazelcast.instance.HazelcastInstanceImpl;
@@ -22,7 +38,7 @@ import java.util.HashMap;
 import java.util.Map;
 
 /**
- * Helper class to be gather JMX related stats for {@link com.hazelcast.management.DefaultTimedMemberStateFactory}
+ * Helper class to be gather JMX related stats for {@link TimedMemberStateFactory}
  */
 final class TimedMemberStateFactoryHelper {
